put E2E tests together
